+++ tf-modules/archive/api.tf
@@ -119,9 +119,15 @@ resource "aws_lambda_function" "api" {
   memory_size = 960
   tags        = var.tags
 
-  vpc_config {
-    subnet_ids         = var.lambda_subnet_ids
-    security_group_ids = var.lambda_subnet_ids == null ? null : [aws_security_group.no_ingress_all_egress[0].id, var.elasticsearch_security_group_id]
+  dynamic "vpc_config" {
+    for_each = length(var.lambda_subnet_ids) == 0 ? [] : [1]
+    content {
+      subnet_ids = var.lambda_subnet_ids
+      security_group_ids = [
+        aws_security_group.no_ingress_all_egress[0].id,
+        var.elasticsearch_security_group_id
+      ]
+    }
   }
 }