+++ tf-modules/ingest/variables.tf
@@ -64,7 +64,7 @@ variable "lambda_processing_role_arn" {
 
 variable "lambda_subnet_ids" {
   type    = list(string)
-  default = null
+  default = []
 }
 
 variable "launchpad_api" {